move defineDynamicBlock into its own file
